requirements.txt
fastapi
uvicorn
bcrypt
python-jose[cryptography]